requirements.txt
# =============================================================================
# Code Learning Coach - Python 依赖
# =============================================================================

# -----------------------------------------------------------------------------
# Core FastAPI & Web Framework
# -----------------------------------------------------------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# -----------------------------------------------------------------------------
# Database (PostgreSQL with async support)
# -----------------------------------------------------------------------------
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# -----------------------------------------------------------------------------
# Configuration & Validation
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
pydantic[email]
pydantic-settings

# -----------------------------------------------------------------------------
# Redis (for caching and Celery broker)
# -----------------------------------------------------------------------------
redis==5.0.1

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
structlog>=23,<25

# -----------------------------------------------------------------------------
# HTTP Client (for external API integrations)
# -----------------------------------------------------------------------------
httpx==0.25.2
httpx-retries>=0.3,<1
tenacity==8.2.3

# -----------------------------------------------------------------------------
# Storage Providers
# -----------------------------------------------------------------------------
aiofiles==23.2.1
anyio==3.7.1
oss2==2.19.0  # Aliyun OSS
boto3==1.34.34  # AWS S3

# -----------------------------------------------------------------------------
# Message Queue (Kafka)
# -----------------------------------------------------------------------------
confluent-kafka>=2.3,<3
aiokafka>=0.10,<1

# -----------------------------------------------------------------------------
# Background Tasks (Celery)
# -----------------------------------------------------------------------------
celery==5.3.6

# -----------------------------------------------------------------------------
# gRPC
# -----------------------------------------------------------------------------
grpcio>=1.58,<2
grpcio-tools>=1.58,<2
grpcio-health-checking>=1.58,<2
googleapis-common-protos>=1.60,<2

# -----------------------------------------------------------------------------
# Internationalization
# -----------------------------------------------------------------------------
Babel>=2.12.0

# =============================================================================
# LangChain Ecosystem (AI/LLM Integration) - v1.0.0+
# =============================================================================

# LangChain Core
langchain>=1.0.0,<2
langchain-core
langchain-community

# LLM Providers
langchain-openai

# Vector Stores (仅 Chroma)
langchain-chroma



# =============================================================================
# LangGraph (Agent Framework)
# =============================================================================
langgraph

# =============================================================================
# Vector Database Clients (仅 Chroma)
# =============================================================================
chromadb>=0.5.0,<0.6

# =============================================================================
# Code Analysis
# =============================================================================

# Tree-sitter for AST parsing
tree-sitter>=0.21.0,<0.23
tree-sitter-python>=0.21.0,<0.23
tree-sitter-javascript>=0.21.0,<0.23
tree-sitter-typescript>=0.21.0,<0.23

# =============================================================================
# Utilities
# =============================================================================
PyYAML>=6.0,<7  # YAML parsing for templates
tiktoken>=0.5.0,<1  # Token counting for OpenAI models
gitpython>=3.1.0,<4  # Git operations (alternative to subprocess)

# =============================================================================
# MinIO Object Storage
# =============================================================================
minio>=7.2.0,<8

# =============================================================================
# Testing
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov>=4.0.0,<5
